// src/types/blocks.ts
export type BlockType = {
  id: any;
  parentId: any;
  block: string;
  layout: any;
  post: any;
  attrs: any;
};

// export type blocksType = {
//     id: any
//     parentId: any
//     block: any
//     layout: any
//     post: any
//     attrs: any
// }[]
// export type blocksType = any
// export type blockType = any